Fix(apimachinery): allow 19-digit quantities to fix math.MaxInt64 parsing and avoid regression on larger values
