+++ staging/src/k8s.io/apimachinery/pkg/api/resource/quantity.go
@@ -320,27 +320,24 @@ func ParseQuantity(str string) (Quantity, error) {
 		if scale >= int32(Nano) {
 			shifted := num + denom
 
-			var value int64
-			value, err := strconv.ParseInt(shifted, 10, 64)
-			if err != nil {
-				return Quantity{}, ErrNumeric
-			}
-			if result, ok := int64Multiply(value, int64(mantissa)); ok {
-				if !positive {
-					result = -result
-				}
-				// if the number is in canonical form, reuse the string
-				switch format {
-				case BinarySI:
-					if exponent%10 == 0 && (value&0x07 != 0) {
-						return Quantity{i: int64Amount{value: result, scale: Scale(scale)}, Format: format, s: str}, nil
+			if value, err := strconv.ParseInt(shifted, 10, 64); err == nil {
+				if result, ok := int64Multiply(value, int64(mantissa)); ok {
+					if !positive {
+						result = -result
 					}
-				default:
-					if scale%3 == 0 && !strings.HasSuffix(shifted, "000") && shifted[0] != '0' {
-						return Quantity{i: int64Amount{value: result, scale: Scale(scale)}, Format: format, s: str}, nil
+					// if the number is in canonical form, reuse the string
+					switch format {
+					case BinarySI:
+						if exponent%10 == 0 && (value&0x07 != 0) {
+							return Quantity{i: int64Amount{value: result, scale: Scale(scale)}, Format: format, s: str}, nil
+						}
+					default:
+						if scale%3 == 0 && !strings.HasSuffix(shifted, "000") && shifted[0] != '0' {
+							return Quantity{i: int64Amount{value: result, scale: Scale(scale)}, Format: format, s: str}, nil
+						}
 					}
+					return Quantity{i: int64Amount{value: result, scale: Scale(scale)}, Format: format}, nil
 				}
-				return Quantity{i: int64Amount{value: result, scale: Scale(scale)}, Format: format}, nil
 			}
 		}
 	}
